+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/dedup/BasePartitionDedupMetadataManager.java
@@ -18,9 +18,15 @@
  */
 package org.apache.pinot.segment.local.dedup;
 
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Preconditions;
+import java.util.Iterator;
 import java.util.List;
+import org.apache.pinot.common.metrics.ServerGauge;
 import org.apache.pinot.common.metrics.ServerMetrics;
+import org.apache.pinot.segment.spi.IndexSegment;
 import org.apache.pinot.spi.config.table.HashFunction;
+import org.apache.pinot.spi.data.readers.PrimaryKey;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -31,14 +37,68 @@ public abstract class BasePartitionDedupMetadataManager implements PartitionDedu
   protected final int _partitionId;
   protected final ServerMetrics _serverMetrics;
   protected final HashFunction _hashFunction;
+  protected final double _metadataTTL;
+  protected final String _metadataTimeColumn;
   protected final Logger _logger;
 
-  protected BasePartitionDedupMetadataManager(String tableNameWithType, int partitionId, DedupContext dedupContext) {
+  protected BasePartitionDedupMetadataManager(String tableNameWithType, int partitionId,
+      DedupContext dedupContext) {
     _tableNameWithType = tableNameWithType;
     _partitionId = partitionId;
     _primaryKeyColumns = dedupContext.getPrimaryKeyColumns();
     _hashFunction = dedupContext.getHashFunction();
     _serverMetrics = dedupContext.getServerMetrics();
+    _metadataTTL = dedupContext.getMetadataTTL() >= 0 ? dedupContext.getMetadataTTL() : 0;
+    _metadataTimeColumn = dedupContext.getMetadataTimeColumn();
+    if (_metadataTTL > 0) {
+      Preconditions.checkArgument(_metadataTimeColumn != null,
+          "When metadataTTL is configured, metadata time column must be configured for dedup enabled table: %s",
+          tableNameWithType);
+    }
     _logger = LoggerFactory.getLogger(tableNameWithType + "-" + partitionId + "-" + getClass().getSimpleName());
   }
+
+  @Override
+  public boolean checkRecordPresentOrUpdate(PrimaryKey pk, IndexSegment indexSegment) {
+    throw new UnsupportedOperationException("checkRecordPresentOrUpdate(PrimaryKey pk, IndexSegment indexSegment) is "
+        + "deprecated!");
+  }
+
+  @Override
+  public void addSegment(IndexSegment segment) {
+    try (DedupUtils.DedupRecordInfoReader dedupRecordInfoReader = new DedupUtils.DedupRecordInfoReader(segment,
+        _primaryKeyColumns, _metadataTimeColumn)) {
+      Iterator<DedupRecordInfo> dedupRecordInfoIterator =
+          DedupUtils.getDedupRecordInfoIterator(dedupRecordInfoReader, segment.getSegmentMetadata().getTotalDocs());
+      addSegment(segment, dedupRecordInfoIterator);
+      int dedupPrimaryKeyCount = removeExpiredPrimaryKeys();
+      _serverMetrics.setValueOfPartitionGauge(_tableNameWithType, _partitionId, ServerGauge.DEDUP_PRIMARY_KEYS_COUNT,
+          dedupPrimaryKeyCount);
+    } catch (Exception e) {
+      throw new RuntimeException(String.format("Caught exception while adding segment: %s of table: %s to "
+          + "ConcurrentMapPartitionDedupMetadataManager", segment.getSegmentName(), _tableNameWithType), e);
+    }
+  }
+
+  @VisibleForTesting
+  protected abstract void addSegment(IndexSegment segment, Iterator<DedupRecordInfo> dedupRecordInfoIterator);
+
+  @Override
+  public void removeSegment(IndexSegment segment) {
+    try (DedupUtils.DedupRecordInfoReader dedupRecordInfoReader = new DedupUtils.DedupRecordInfoReader(segment,
+        _primaryKeyColumns, _metadataTimeColumn)) {
+      Iterator<DedupRecordInfo> dedupRecordInfoIterator =
+          DedupUtils.getDedupRecordInfoIterator(dedupRecordInfoReader, segment.getSegmentMetadata().getTotalDocs());
+      removeSegment(segment, dedupRecordInfoIterator);
+      int dedupPrimaryKeyCount = removeExpiredPrimaryKeys();
+      _serverMetrics.setValueOfPartitionGauge(_tableNameWithType, _partitionId, ServerGauge.DEDUP_PRIMARY_KEYS_COUNT,
+          dedupPrimaryKeyCount);
+    } catch (Exception e) {
+      throw new RuntimeException(String.format("Caught exception while removing segment: %s of table: %s from "
+          + "ConcurrentMapPartitionDedupMetadataManager", segment.getSegmentName(), _tableNameWithType), e);
+    }
+  }
+
+  @VisibleForTesting
+  protected abstract void removeSegment(IndexSegment segment, Iterator<DedupRecordInfo> dedupRecordInfoIterator);
 }
